Rate Limiting feature name should match folder name
