--- src/Ocelot/RateLimit/IRateLimitCounterHandler.cs
@@ -1,13 +1,0 @@
-﻿namespace Ocelot.RateLimit
-{
-    public interface IRateLimitCounterHandler
-    {
-        bool Exists(string id);
-
-        RateLimitCounter? Get(string id);
-
-        void Remove(string id);
-
-        void Set(string id, RateLimitCounter counter, TimeSpan expirationTime);
-    }
-}
